Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
elevenlabs==0.2.26
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
